Additional ZeroClipboard Removal
