+++ app/addons/components/components/copy.js
@@ -44,6 +44,24 @@ export class Copy extends React.Component {
   }
 
   render () {
+    if (this.props.displayType === 'input') {
+      return (
+        <p>
+          <input
+            type="text"
+            className="input-xxlarge text-field-to-copy"
+            readOnly
+            value={this.props.text} />
+          <button
+            className="btn copy-button clipboard-copy-element"
+            data-clipboard-text={this.props.text}
+            title={this.props.title}
+          >
+            {this.props.textDisplay}
+          </button>
+        </p>
+      );
+    }
     return (
       <button className="copy clipboard-copy-element"
               title={this.props.title}
@@ -64,6 +82,70 @@ Copy.defaultProps = {
 
 Copy.propTypes = {
   text: React.PropTypes.string.isRequired,
-  displayType: React.PropTypes.oneOf(['icon', 'text']),
+  displayType: React.PropTypes.oneOf(['icon', 'text', 'input']),
   onClipboardClick: React.PropTypes.func.isRequired
 };
+
+
+
+/*
+
+var ClipboardWithTextField = React.createClass({
+  propTypes: {
+    onClipBoardClick: React.PropTypes.func.isRequired,
+    textToCopy: React.PropTypes.string.isRequired,
+    uniqueKey: React.PropTypes.string.isRequired,
+    showCopyIcon: React.PropTypes.bool
+  },
+
+  getDefaultProps: function () {
+    return {
+      showCopyIcon: true,
+      text: 'Copy'
+    };
+  },
+
+  componentWillMount: function () {
+    ZeroClipboard.config({ swfPath: getZeroClipboardSwfPath() });
+  },
+
+  componentDidMount: function () {
+    var el = ReactDOM.findDOMNode(this.refs["copy-text-" + this.props.uniqueKey]);
+    this.clipboard = new ZeroClipboard(el);
+    this.clipboard.on('ready', function () {
+      this.clipboard.on('copy', function () {
+        this.props.onClipBoardClick();
+      }.bind(this));
+    }.bind(this));
+  },
+
+  getCopyIcon: function () {
+    if (!this.props.showCopyIcon) {
+      return null;
+    }
+    return (<i className="fontawesome icon-paste"></i>);
+  },
+
+  render: function () {
+    return (
+      <p key={this.props.uniqueKey}>
+        <input
+          type="text"
+          className="input-xxlarge text-field-to-copy"
+          readOnly
+          value={this.props.textToCopy} />
+        <a
+          id={"copy-text-" + this.props.uniqueKey}
+          className="btn copy-button clipboard-copy-element"
+          data-clipboard-text={this.props.textToCopy}
+          data-bypass="true"
+          ref={"copy-text-" + this.props.uniqueKey}
+          title="Copy to clipboard"
+        >
+          {this.getCopyIcon()} {this.props.text}
+        </a>
+      </p>
+    );
+  }
+});
+*/
