+++ app/addons/fauxton/assets/less/components.less
@@ -23,10 +23,6 @@
       }
     }
 
-    .zeroclipboard-is-hover {
-      background-color: #cbcbcb;
-    }
-
     .icon-question-sign {
       margin-left: 3px;
     }
